requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
websockets==13.1
pydantic-settings==2.5.2
openai==1.65.2
PyPDF2==3.0.1
jinja2==3.1.4
python-multipart==0.0.12
python-dotenv==1.0.1